lsm
Enum Capability

#[repr(usize)]
pub enum Capability {

    Volumes = 20,
    VolumeCreate = 21,
    VolumeResize = 22,
    VolumeReplicate = 23,
    VolumeReplicateClone = 24,
    VolumeReplicateCopy = 25,
    VolumeReplicateMirrorAsync = 26,
    VolumeReplicateMirrorSync = 27,
    VolumeRepRangeBlockSize = 28,
    VolumeRepRange = 29,
    VolumeRepRangeClone = 30,
    VolumeRepRangeCopy = 31,
    VolumeDelete = 33,
    VolumeEnable = 34,
    VolumeDisable = 35,
    VolumeMask = 36,
    VolumeUnmask = 37,
    AccessGroups = 38,
    AccessGroupCreateWwpn = 39,
    AccessGroupDelete = 40,
    AccessGroupInitAddWwpn = 41,
    AccessGroupInitDel = 42,
    VolsMaskedToAg = 43,
    AgsGrantedToVol = 44,
    VolHasChildDep = 45,
    VolChildDepRm = 46,
    AccessGroupCreateIscsiIqn = 47,
    AccessGroupInitAddIscsiIqn = 48,
    IscsiChapAuthSet = 53,
    VolRaidInfo = 54,
    VolumeThin = 55,
    Batteries = 56,
    VolCacheInfo = 57,
    VolPhyDiskCacheSet = 58,
    VolPhysicalDiskCacheSetSystemLevel = 59,
    VolWriteCacheSetEnable = 60,
    VolWriteCacheSetAuto = 61,
    VolWriteCacheSetDisabled = 62,
    VolWriteCacheSetImpactRead = 63,
    VolWriteCacheSetWbImpactOther = 64,
    VolReadCacheSet = 65,
    VolReadCacheSetImpactWrite = 66,
    Fs = 100,
    FsDelete = 101,
    FsResize = 102,
    FsCreate = 103,
    FsClone = 104,
    FsFileClone = 105,
    FsSnapshots = 106,
    FsSnapshotCreate = 107,
    FsSnapshotDelete = 109,
    FsSnapshotRestore = 110,
    FsSnapshotRestoreSpecificFiles = 111,
    FsHasChildDep = 112,
    FsChildDepRm = 113,
    FsChildDepRmSpecificFiles = 114,
    NfsExportAuthTypeList = 120,
    NfsExports = 121,
    FsExport = 122,
    FsUnexport = 123,
    FsExportCustomPath = 124,
    SysReadCachePctSet = 158,
    SysReadCachePctGet = 159,
    SysFwVersionGet = 160,
    SysModeGet = 161,
    DiskLocation = 163,
    DiskRpm = 164,
    DiskLinkType = 165,
    VolumeLed = 171,
    TargetPorts = 216,
    Disks = 220,
    PoolMemberInfo = 221,
    VolumeRaidCreate = 222,
    DiskVpd83Get = 223,

}

Represent a capability supported by specific system.

Variants

Volumes = 20

Support Client::volumes().

VolumeCreate = 21

Support Client::volume_create().

VolumeResize = 22

Support Client::volume_resize().

VolumeReplicate = 23

Support Client::volume_replicate().

VolumeReplicateClone = 24

Support Client::volume_replicate() with VolumeReplicateType::Clone.

VolumeReplicateCopy = 25

Support Client::volume_replicate() with VolumeReplicateType::Copy.

VolumeReplicateMirrorAsync = 26

Support Client::volume_replicate() with VolumeReplicateType::MirrorAsync.

VolumeReplicateMirrorSync = 27

Support Client::volume_replicate() with VolumeReplicateType::MirrorSync.

VolumeRepRangeBlockSize = 28

Support Client::volume_rep_range_blk_size().

VolumeRepRange = 29

Support Client::volume_rep_range().

VolumeRepRangeClone = 30

Support Client::volume_rep_range() with VolumeReplicateType::Clone.

VolumeRepRangeCopy = 31

Support Client::volume_rep_range() with VolumeReplicateType::Copy.

VolumeDelete = 33

Support Client::volume_delete().

VolumeEnable = 34

Support Client::volume_enable().

VolumeDisable = 35

Support Client::volume_disable().

VolumeMask = 36

Support Client::volume_mask().

VolumeUnmask = 37

Support Client::volume_unmask().

AccessGroups = 38

Support Client::access_groups().

AccessGroupCreateWwpn = 39

Support Client::access_group_create() with InitiatorType::Wwpn.

AccessGroupDelete = 40

Support Client::access_group_delete().

AccessGroupInitAddWwpn = 41

Support Client::access_group_init_add() with InitiatorType::Wwpn.

AccessGroupInitDel = 42

Support Client::access_group_init_del().

VolsMaskedToAg = 43

Support Client::vols_masked_to_ag().

AgsGrantedToVol = 44

Support Client::ags_granted_to_vol().

VolHasChildDep = 45

Support Client::vol_has_child_dep().

VolChildDepRm = 46

Support Client::vol_child_dep_rm().

AccessGroupCreateIscsiIqn = 47

Support Client::access_group_create() with InitiatorType::IscsiIqn.

AccessGroupInitAddIscsiIqn = 48

Support Client::access_group_init_add() with InitiatorType::IscsiIqn.

IscsiChapAuthSet = 53

Support Client::iscsi_chap_auth_set().

VolRaidInfo = 54

Support Client::vol_raid_info().

VolumeThin = 55

Support Client::volume_crate() with thinp=VolumeCreateArgThinP::Thin argument.

Batteries = 56

Support Client::batteries().

VolCacheInfo = 57

Support Client::vol_cache_info().

VolPhyDiskCacheSet = 58

Support Client::vol_phy_disk_cache_set().

VolPhysicalDiskCacheSetSystemLevel = 59

Indicate the Client::vol_phy_disk_cache_set() will change system settings which are effective on all volumes in this storage system. For example, on HPE SmartArray, the physical disk cache setting is a controller level setting.

VolWriteCacheSetEnable = 60

Support Client::vol_write_cache_set() with wcp=Cache::Enabled.

VolWriteCacheSetAuto = 61

Support Client::vol_write_cache_set() with wcp=Cache::Auto.

VolWriteCacheSetDisabled = 62

Support Client::vol_write_cache_set() with wcp=Cache::Disabled.

VolWriteCacheSetImpactRead = 63

Indicate the Client::vol_write_cache_set() might also impact read cache policy.

VolWriteCacheSetWbImpactOther = 64

Indicate the Client::vol_write_cache_set() with wcp=Cache::Enabled might impact other volumes in the same system.

VolReadCacheSet = 65

Support Client::vol_read_cache_set().

VolReadCacheSetImpactWrite = 66

Indicate the Client::vol_read_cache_set() might also impact write cache policy.

Fs = 100

Support Client::fs().

FsDelete = 101

Support Client::fs_delete().

FsResize = 102

Support Client::fs_resize().

FsCreate = 103

Support Client::fs_create().

FsClone = 104

Support Client::fs_clone().

FsFileClone = 105

Support Client::fs_file_clone().

FsSnapshots = 106

Support Client::fs_snapshots().

FsSnapshotCreate = 107

Support Client::fs_snapshot_create().

FsSnapshotDelete = 109

Support Client::fs_snapshot_delete().

FsSnapshotRestore = 110

Support Client::fs_snapshot_restore().

FsSnapshotRestoreSpecificFiles = 111

Support Client::fs_snapshot_restore() with files argument.

FsHasChildDep = 112

Support Client::fs_has_child_dep().

FsChildDepRm = 113

Support Client::fs_child_dep_rm().

FsChildDepRmSpecificFiles = 114

Support Client::fs_child_dep_rm() with files argument.

NfsExportAuthTypeList = 120

Support Client:::nfs_exp_auth_type_list().

NfsExports = 121

Support Client::nfs_exports().

FsExport = 122

Support Client::fs_export().

FsUnexport = 123

Support Client::fs_unexport().

FsExportCustomPath = 124

Support Client::fs_export() with export_path argument.

SysReadCachePctSet = 158

Support Client::system_read_cache_pct_set()

SysReadCachePctGet = 159

Support Client::systems() with valid read_cache_pct property.

SysFwVersionGet = 160

Support Client::systems() with valid fw_version property.

SysModeGet = 161

Support Client::systems() with valid mode property.

DiskLocation = 163

Support Client::disks() with valid location property.

DiskRpm = 164

Support Client::disks() with valid rpm property.

DiskLinkType = 165

Support Client::disks() with valid link_type property.

VolumeLed = 171

Support Client::vol_ident_led_on() and Client::vol_ident_led_off().

TargetPorts = 216

Support Client::target_ports().

Disks = 220

Support Client::disks().

PoolMemberInfo = 221

Support Client::pool_member_info().

VolumeRaidCreate = 222

Support Client::vol_raid_create_cap_get() and Client::vol_raid_create().

DiskVpd83Get = 223

Support Client::disks() with valid vpd83 property.