Enum lsm::VolumeReplicateType

pub enum VolumeReplicateType {
    Unknown,
    Clone,
    Copy,
    MirrorSync,
    MirrorAsync,
}

Represent a volume replication type.

Variants
Unknown

Plugin failed to detect volume replication type.

Clone

Point in time read writeable space efficient copy of data. Also know as read writeable snapshot.

Copy

Full bitwise copy of the data (occupies full space).

MirrorSync

I/O will be blocked until I/O reached both source and target storage systems. There will be no data difference between source and target storage systems.

MirrorAsync

I/O will be blocked until I/O reached source storage systems. The source storage system will use copy the changes data to target system in a predefined interval. There will be a small data differences between source and target.