lsm
Module local_disk

Structs

LedSlot

Information about specific slot

LedSlots

Opaque type for interacting with LED slots functionality.

Enums

HealthStatus

Possible health status of any disk.

LinkType

Possible values of link type.

Rpm

Possible values for disk RPM.

Constants

LED_STATUS_FAULT_OFF

Fault LED is off

LED_STATUS_FAULT_ON

Fault LED is on

LED_STATUS_FAULT_UNKNOWN

Fault LED is unknown

LED_STATUS_IDENT_OFF

Identification LED is off

LED_STATUS_IDENT_ON

Identification LED is on

LED_STATUS_IDENT_UNKNOWN

Identification LED is unknown

LED_STATUS_UNKNOWN

LED has unknown status

Functions

fault_led_off

Turn the fault LED off for specified disk path.

fault_led_on

Turn the fault LED on for specified disk path.

health_get

Query the health status of the specified disk path.

ident_led_off

Turn the identification LED off for specified disk path.

ident_led_on

led_status_get

Retrieve current state of LED for specified disk path.

link_speed_get

Query the current negotiated disk link speed in Mbps

link_type_get

Query disk link type.

list

rpm_get

Query disk rotation speed.

serial_num_get

Query serial number of specified disk path

vpd83_get

Query scsi VPD 0x83 NAA ID.

vpd83_search

Searches disk by supplied SCSI VPD 0x83 page NAA type ID.