lsm
Struct System

pub struct System {
    pub id: String,
    pub name: String,
    pub status: u32,
    pub status_info: String,
    pub fw_version: String,
    pub read_cache_pct: i8,
    pub mode: SystemMode,
    /* private fields */
}

Represent a storage system. Examples:

A hardware RAID card, LSI MegaRAID

A storage area network (SAN), e.g. EMC VNX, NetApp Filer

A software solution running on commodity hardware, targetd, Nexenta

A Linux system running NFS service

Fields
id: String

Identifier.

name: String

Human friendly name.

status: u32

System status stored in bitmap. Valid status value are:

System::STATUS_UNKNOWN

System::STATUS_OK

System::STATUS_ERROR

System::STATUS_DEGRADED

System::STATUS_PREDICTIVE_FAILURE

System::STATUS_OTHER

if (s.status & System::STATUS_OK) == 0 {
    println!("System {}/{} is not healthy", s.name, s.id);
}

status_info: String

Additional message for status.

fw_version: String

Firmware version.

read_cache_pct: i8

Read cache percentage of the system. Valid values are:

>0 and < 100 means only a part of whole cache are used for read.

0 means no read cache.

100 means all cache are used for read.

System::READ_CACHE_PCT_NO_SUPPORT means no support.

System::READ_CACHE_PCT_UNKNOWN means plugin failed to detect this value.

mode: SystemMode

System mode, currently only supports hardware RAID cards.

Implementations

impl System

pub const READ_CACHE_PCT_NO_SUPPORT: i8 = -2i8

Plugin does not support querying read cache percentage.

pub const READ_CACHE_PCT_UNKNOWN: i8 = -1i8

Plugin failed to query read cache percentage.

pub const STATUS_UNKNOWN: u32 = 1u32

Plugin failed to query system status.

pub const STATUS_OK: u32 = 2u32

System is up and healthy.

pub const STATUS_ERROR: u32 = 4u32

System is in error state.

pub const STATUS_DEGRADED: u32 = 8u32

System is degraded.

pub const STATUS_PREDICTIVE_FAILURE: u32 = 16u32

System has potential failure.

pub const STATUS_OTHER: u32 = 32u32

Vendor specific status.