Enum lsm::Capability

#[repr(usize)]

pub enum Capability {
    Volumes,
    VolumeCreate,
    VolumeResize,
    VolumeReplicate,
    VolumeReplicateClone,
    VolumeReplicateCopy,
    VolumeReplicateMirrorAsync,
    VolumeReplicateMirrorSync,
    VolumeRepRangeBlockSize,
    VolumeRepRange,
    VolumeRepRangeClone,
    VolumeRepRangeCopy,
    VolumeDelete,
    VolumeEnable,
    VolumeDisable,
    VolumeMask,
    VolumeUnmask,
    AccessGroups,
    AccessGroupCreateWwpn,
    AccessGroupDelete,
    AccessGroupInitAddWwpn,
    AccessGroupInitDel,
    VolsMaskedToAg,
    AgsGrantedToVol,
    VolHasChildDep,
    VolChildDepRm,
    AccessGroupCreateIscsiIqn,
    AccessGroupInitAddIscsiIqn,
    IscsiChapAuthSet,
    VolRaidInfo,
    VolumeThin,
    Batteries,
    VolCacheInfo,
    VolPhyDiskCacheSet,
    VolPhysicalDiskCacheSetSystemLevel,
    VolWriteCacheSetEnable,
    VolWriteCacheSetAuto,
    VolWriteCacheSetDisabled,
    VolWriteCacheSetImpactRead,
    VolWriteCacheSetWbImpactOther,
    VolReadCacheSet,
    VolReadCacheSetImpactWrite,
    Fs,
    FsDelete,
    FsResize,
    FsCreate,
    FsClone,
    FsFileClone,
    FsSnapshots,
    FsSnapshotCreate,
    FsSnapshotDelete,
    FsSnapshotRestore,
    FsSnapshotRestoreSpecificFiles,
    FsHasChildDep,
    FsChildDepRm,
    FsChildDepRmSpecificFiles,
    NfsExportAuthTypeList,
    NfsExports,
    FsExport,
    FsUnexport,
    FsExportCustomPath,
    SysReadCachePctSet,
    SysReadCachePctGet,
    SysFwVersionGet,
    SysModeGet,
    DiskLocation,
    DiskRpm,
    DiskLinkType,
    VolumeLed,
    TargetPorts,
    Disks,
    PoolMemberInfo,
    VolumeRaidCreate,
    DiskVpd83Get,
}

Represent a capability supported by specific system.

Variants
Volumes

Support Client::volumes().

VolumeCreate

Support Client::volume_create().

VolumeResize

Support Client::volume_resize().

VolumeReplicate

Support Client::volume_replicate().

VolumeReplicateClone

Support Client::volume_replicate() with VolumeReplicateType::Clone.

VolumeReplicateCopy

Support Client::volume_replicate() with VolumeReplicateType::Copy.

VolumeReplicateMirrorAsync

Support Client::volume_replicate() with VolumeReplicateType::MirrorAsync.

VolumeReplicateMirrorSync

Support Client::volume_replicate() with VolumeReplicateType::MirrorSync.

VolumeRepRangeBlockSize

Support Client::volume_rep_range_blk_size().

VolumeRepRange

Support Client::volume_rep_range().

VolumeRepRangeClone

Support Client::volume_rep_range() with VolumeReplicateType::Clone.

VolumeRepRangeCopy

Support Client::volume_rep_range() with VolumeReplicateType::Copy.

VolumeDelete

Support Client::volume_delete().

VolumeEnable

Support Client::volume_enable().

VolumeDisable

Support Client::volume_disable().

VolumeMask

Support Client::volume_mask().

VolumeUnmask

Support Client::volume_unmask().

AccessGroups

Support Client::access_groups().

AccessGroupCreateWwpn

Support Client::access_group_create() with InitiatorType::Wwpn.

AccessGroupDelete

Support Client::access_group_delete().

AccessGroupInitAddWwpn

Support Client::access_group_init_add() with InitiatorType::Wwpn.

AccessGroupInitDel

Support Client::access_group_init_del().

VolsMaskedToAg

Support Client::vols_masked_to_ag().

AgsGrantedToVol

Support Client::ags_granted_to_vol().

VolHasChildDep

Support Client::vol_has_child_dep().

VolChildDepRm

Support Client::vol_child_dep_rm().

AccessGroupCreateIscsiIqn

Support Client::access_group_create() with InitiatorType::IscsiIqn.

AccessGroupInitAddIscsiIqn

Support Client::access_group_init_add() with InitiatorType::IscsiIqn.

IscsiChapAuthSet

Support Client::iscsi_chap_auth_set().

VolRaidInfo

Support Client::vol_raid_info().

VolumeThin

Support Client::volume_crate() with thinp=VolumeCreateArgThinP::Thin argument.

Batteries

Support Client::batteries().

VolCacheInfo

Support Client::vol_cache_info().

VolPhyDiskCacheSet

Support Client::vol_phy_disk_cache_set().

VolPhysicalDiskCacheSetSystemLevel

Indicate the Client::vol_phy_disk_cache_set() will change system settings which are effective on all volumes in this storage system. For example, on HPE SmartArray, the physical disk cache setting is a controller level setting.

VolWriteCacheSetEnable

Support Client::vol_write_cache_set() with wcp=Cache::Enabled.

VolWriteCacheSetAuto

Support Client::vol_write_cache_set() with wcp=Cache::Auto.

VolWriteCacheSetDisabled

Support Client::vol_write_cache_set() with wcp=Cache::Disabled.

VolWriteCacheSetImpactRead

Indicate the Client::vol_write_cache_set() might also impact read cache policy.

VolWriteCacheSetWbImpactOther

Indicate the Client::vol_write_cache_set() with wcp=Cache::Enabled might impact other volumes in the same system.

VolReadCacheSet

Support Client::vol_read_cache_set().

VolReadCacheSetImpactWrite

Indicate the Client::vol_read_cache_set() might also impact write cache policy.

Fs

Support Client::fs().

FsDelete

Support Client::fs_delete().

FsResize

Support Client::fs_resize().

FsCreate

Support Client::fs_create().

FsClone

Support Client::fs_clone().

FsFileClone

Support Client::fs_file_clone().

FsSnapshots

Support Client::fs_snapshots().

FsSnapshotCreate

Support Client::fs_snapshot_create().

FsSnapshotDelete

Support Client::fs_snapshot_delete().

FsSnapshotRestore

Support Client::fs_snapshot_restore().

FsSnapshotRestoreSpecificFiles

Support Client::fs_snapshot_restore() with files arugment.

FsHasChildDep

Support Client::fs_has_child_dep().

FsChildDepRm

Support Client::fs_child_dep_rm().

FsChildDepRmSpecificFiles

Support Client::fs_child_dep_rm() with files argument.

NfsExportAuthTypeList

Support Client:::nfs_exp_auth_type_list().

NfsExports

Support Client::nfs_exports().

FsExport

Support Client::fs_export().

FsUnexport

Support Client::fs_unexport().

FsExportCustomPath

Support Client::fs_export() with export_path argument.

SysReadCachePctSet

Support Client::system_read_cache_pct_set()

SysReadCachePctGet

Support Client::systems() with valid read_cache_pct property.

SysFwVersionGet

Support Client::systems() with valid fw_version property.

SysModeGet

Support Client::systems() with valid mode property.

DiskLocation

Support Client::disks() with valid location property.

DiskRpm

Support Client::disks() with valid rpm property.

DiskLinkType

Support Client::disks() with valid link_type property.

VolumeLed

Support Client::vol_ident_led_on() and Client::vol_ident_led_off().

TargetPorts

Support Client::target_ports().

Disks

Support Client::disks().

PoolMemberInfo

Support Client::pool_member_info().

VolumeRaidCreate

Support Client::vol_raid_create_cap_get() and Client::vol_raid_create().

DiskVpd83Get

Support Client::disks() with valid vpd83 property.