Struct lsm::TargetPort

pub struct TargetPort {
    pub id: String,
    pub port_type: PortType,
    pub service_address: String,
    pub network_address: String,
    pub physical_address: String,
    pub physical_name: String,
    pub system_id: String,
    // some fields omitted
}

Represent a target port which is the front-end port of storage system which storage user/client connect to and get storage service from.

Fields
id: String

Identifier.

port_type: PortType

Type of port.

service_address: String

The address used by upper layer like FC and iSCSI:

FC and FCoE: WWPN

iSCSI: IQN The string is in lower case, split with : every two digits if WWPN.

network_address: String

The address used by network layer like FC and TCP/IP:

FC/FCoE: WWPN

iSCSI: IPv4:Port or [IPv6]:Port The string is in lower case, split with : every two digits if WWPN.

physical_address: String

The address used by physical layer like FC-0 and MAC:

FC and FCoE : WWPN

iSCSI: MAC The string is in Lower case, split with : every two digits.

physical_name: String

The name of physical port. Administrator could use this name to locate the port on storage system. E.g. 'eth0'

system_id: String

Identifier of owner system.